styling and database work
put in styling for pages. Pretty nav bar. Footer is at the bottom with a bg. Adjusted header sizing. Added a font.

Also brought in files that were previously made. A photo of the database schema (made with draw sql), and an sql file to set up the db tables